requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Image Processing
opencv-python-headless==4.8.1.78
Pillow==10.1.0
scikit-image==0.22.0
scipy==1.11.4
numpy==1.26.2

# ML/AI
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
face-recognition==1.3.0

# MRZ Extraction
readmrz
passport-mrz-extractor
pytesseract==0.3.10

# Utilities
psutil==5.9.6
nest-asyncio==1.5.8
requests==2.31.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2